Experimenting with perf ideas
I'm running a small experiment to see if setImmediate helps free
up the thread when GET requests come in. We seem to have a very
slow time caching data, which negatively impacts interactions of
any other user request (POST/GET, etc).

This is a small experiment to see if this helps that, and also to
see how much it delays the existing page loads.

Ideally I'll through a db in to clean this all up... someday.....