Fetch Redis in smaller chunks.
So I had a hunch, that the slowest part of the app was the app was
fetching data from redis. It turns out that the redis client's
lrange function is blocking, and we were fetching a large amount
of data.

I wanted to unblock the thread, so I changed the `lrange 0, -1` into
small chunks, `lrange 0 1000`, `lrange 1000 2000`, ...

I was expecting to at the very least unblock GET/POST requests.
I was not expecting a 55x speed improvement. Fetching everything
and concatting it to an array dramatically improves performance.